requirements.txt
tqdm
pycairo==1.18.1
msgpack_numpy==0.4.4.2
future==0.17.1
parse==1.11.1
svgpathtools==1.3.3
easydict==1.9
tabulate==0.8.3
transforms3d==0.3.1
numpy==1.16.2
matplotlib==3.1.1
python_gflags==3.1.2
msgpack_python==0.5.6
PyYAML==5.1.2
pyzmq==18.1.0
